Move Sinkhorn jitting tests, add chex as dep
