+++ setup.cfg
@@ -54,6 +54,7 @@ test =
     pytest-xdist
     pytest-memray
     pytest-cov
+    chex
     absl-py
 docs =
     sphinx>=4.0
