--- tests/core/sinkhorn_jit_test.py
@@ -1,117 +1,0 @@
-# Copyright 2022 Google LLC.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Jitting test for Sinkhorn."""
-import functools
-from typing import Callable
-
-import chex
-import jax
-import jax.numpy as jnp
-import pytest
-
-from ott.core import sinkhorn
-from ott.geometry import geometry
-
-non_jitted_sinkhorn = functools.partial(sinkhorn.sinkhorn, jit=False)
-
-
-class TestSinkhornJIT:
-  """Check jitted and non jit match for Sinkhorn, and that everything jits."""
-
-  @pytest.fixture(autouse=True)
-  def initialize(self, rng: jnp.ndarray):
-    self.dim = 3
-    self.n = 10
-    self.m = 11
-    self.rng, *rngs = jax.random.split(rng, 10)
-    self.rngs = rngs
-    self.x = jax.random.uniform(rngs[0], (self.n, self.dim))
-    self.y = jax.random.uniform(rngs[1], (self.m, self.dim))
-    a = jax.random.uniform(rngs[2], (self.n,)) + .1
-    b = jax.random.uniform(rngs[3], (self.m,)) + .1
-
-    self.a = a / jnp.sum(a)
-    self.b = b / jnp.sum(b)
-    self.epsilon = 0.05
-    self.geometry = geometry.Geometry(
-        cost_matrix=(
-            jnp.sum(self.x ** 2, axis=1)[:, jnp.newaxis] +
-            jnp.sum(self.y ** 2, axis=1)[jnp.newaxis, :] -
-            2 * jnp.dot(self.x, self.y.T)
-        ),
-        epsilon=self.epsilon
-    )
-
-  @pytest.mark.fast
-  def test_jit_vs_non_jit_fwd(self):
-
-    def assert_output_close(x: jnp.ndarray, y: jnp.ndarray) -> None:
-      """Asserst SinkhornOutputs are close."""
-      x = tuple(a for a in x if (a is not None and isinstance(a, jnp.ndarray)))
-      y = tuple(a for a in y if (a is not None and isinstance(a, jnp.ndarray)))
-      return chex.assert_tree_all_close(x, y, atol=1e-6, rtol=0)
-
-    def f(
-        g: geometry.Geometry, a: jnp.ndarray, b: jnp.ndarray
-    ) -> sinkhorn.SinkhornOutput:
-      return non_jitted_sinkhorn(g, a, b)
-
-    jitted_result = sinkhorn.sinkhorn(self.geometry, self.a, self.b)
-    non_jitted_result = non_jitted_sinkhorn(self.geometry, self.a, self.b)
-    user_jitted_result = jax.jit(f)(self.geometry, self.a, self.b)
-
-    assert_output_close(jitted_result, non_jitted_result)
-    assert_output_close(jitted_result, user_jitted_result)
-
-  @pytest.mark.parametrize("implicit", [False, True])
-  def test_jit_vs_non_jit_bwd(self, implicit: bool):
-
-    def loss(
-        a: jnp.ndarray, x: jnp.ndarray, fun: Callable[[...],
-                                                      sinkhorn.SinkhornOutput]
-    ):
-      out = fun(
-          geometry.Geometry(
-              cost_matrix=(
-                  jnp.sum(x ** 2, axis=1)[:, jnp.newaxis] +
-                  jnp.sum(self.y ** 2, axis=1)[jnp.newaxis, :] -
-                  2 * jnp.dot(x, self.y.T)
-              ),
-              epsilon=self.epsilon
-          ),
-          a=a,
-          b=self.b,
-          tau_a=0.94,
-          tau_b=0.97,
-          threshold=1e-4,
-          lse_mode=True,
-          implicit_differentiation=implicit
-      )
-      return out.reg_ot_cost
-
-    def value_and_grad(a: jnp.ndarray, x: jnp.ndarray):
-      return jax.value_and_grad(loss)(a, x, non_jitted_sinkhorn)
-
-    jitted_loss, jitted_grad = jax.value_and_grad(loss)(
-        self.a, self.x, sinkhorn.sinkhorn
-    )
-    non_jitted_loss, non_jitted_grad = jax.value_and_grad(loss)(
-        self.a, self.x, non_jitted_sinkhorn
-    )
-
-    user_jitted_loss, user_jitted_grad = jax.jit(value_and_grad)(self.a, self.x)
-    chex.assert_tree_all_close(jitted_loss, non_jitted_loss, atol=1e-6, rtol=0)
-    chex.assert_tree_all_close(jitted_grad, non_jitted_grad, atol=1e-6, rtol=0)
-    chex.assert_tree_all_close(user_jitted_loss, jitted_loss, atol=1e-6, rtol=0)
-    chex.assert_tree_all_close(user_jitted_grad, jitted_grad, atol=1e-6, rtol=0)
